requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
Pillow>=9.0.0
rawpy>=0.18.0
numpy>=1.24.0